feat: cleaning main function
