+++ SimpleTodoListApp/functions/addTask.py
@@ -1,4 +1,5 @@
-def addTask(newTask: str, taskList: list):
+def addTask(taskList: list):
+    newTask = input("Enter the task: ")
     if newTask not in taskList:
         taskList.append(newTask)
         print(f"'{newTask}' has been added to the list")
